requirements.txt
torch>=2.0.1
torchvision>=0.15.0
torchaudio>=2.0.0
wandb>=0.15.0
pyyaml>=6.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
timm==0.9.16
transformers>=4.20.0
grad-cam>=1.4.0
albumentations>=1.3.0
pandas>=1.3.0
pytest>=7.0.0
jupyter>=1.0.0
torchsummary>=1.5.1
huggingface-hub==0.22.2

# Development dependencies (optional)
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950
